Typescript vector data
