--- src/state/postgres.js
@@ -1,19 +1,0 @@
-class Postgres {
-  constructor({
-    username,
-    password,
-    port,
-    url,
-    databaseName,
-    scrapeInterval,
-  }) {
-    this.username = username
-    this.password = password
-    this.port = port
-    this.url = url
-    this.databaseName = databaseName
-    this.scrapeInterval = scrapeInterval
-  }
-}
-
-module.exports = Postgres
