--- src/state/vector.js
@@ -1,11 +1,0 @@
-const Postgres = require('./postgres')
-
-class Vector {
-  constructor({
-    postgres,
-  }) {
-    this.Postgres = new Postgres(postgres)
-  }
-}
-
-module.exports = Vector
